Added Picozed fpgadrive PCIe project
This patch adds the picozed fpgadrive PCIe project.  We also reorganize the toplevel folder structure and
hence updated README.md

The idea is to implement the design on the zedboard, that also instantiates a Xilinx PCIe subsystem, for testing purposes.
Note that for instantiation purposes, we will map the PCIe endpoints to meaningless pins on the FPGA.